app/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

Flask==2.2.5
gunicorn==20.0.4
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy>=1.22,<2.0
pandas>=1.4,<2.0
Pillow>=8.0.1,<10.0.0
requests>=2.25.0
scikit-learn>=1.2,<1.4

# PyTorch (CPU) - avoid hard-pinning ancient cp36 wheels
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2